Fix pep8 issues and add apache header to subunit2html
This commit fixes pep8 issues that were added when subunit2html was
copied into the repo. As part of this the subunit2html was missing a
license header for the additions made to it on top of the original
HTMLTestRunner script it was based on. The addition of the copyright
header mirrors project-config change:
I6fc16d316012fd3e1c196f74df25d725a310f6dc

Change-Id: Ieb42d5baddb3e1446fcd50197136f605324323aa