Add support for having comments in the exclude file
This commit adds the support to have comments in the exclude file.
After this commit anything after a '#' will be treated as a comment
and ignored for matching. A new option --print-exclude is added which
will print out what is being skipped by the exclude file with any
comments in the file.

Change-Id: Ia357730f918e0a57cf2ac51cd8195e2721667511