Disable printing percent change on run time by default
This commit disables printing the percent change in run time by
default. This was a cool experiment, but isn't really useful in
practice. Especially in the gate, things are so noisy and there
is so much variance looking at the change for a single doesn't
actually mean anything. It makes more sense to make this feature
opt-in because it's not generally useful.

Change-Id: Iecb153452edfe1d7b55757d022ae0331ac563b35