Add TODO entry for moving away from subprocess in ostestr
This commit adds a new entry to the os-testr TODO file to indicate the
long term goal for migrating away from using subprocess in ostestr.
Everything ostestr uses is a python project so we shouldn't need to
call out for everything, it was just faster to bootstrap the project
this way, but we should probably move away from doing this.

Change-Id: I3b2e54d72514ad92079a893fd7a835e6d446d4bc