Convert file names to regular expressions
Add the capability to the --no-discover option to convert file
names into regular sexpressions.  Also add the --path option that
converts a file or directory to a regular expression.  Create a
mutually exclusive argparse group for --regex, --path, and
--no-discover.

This change will allow the user to specify a file name instead of a
regular expression to match a particular set of tests e.g:

    tox -e py27 -- --path os_testr/tests/test_os_testr.py

Will run os_testr.tests.test_os_testr but you can use tab complete
to generate the name.

Change-Id: Ibfca2bc023aed44b1b87a0444559ab2a00303a70