Use test_to_run var in no-discover
The use of -n or --pdb was broken if a user
specified the test by path.  The problem is
that even though in the case of -n we parse
out path and convert to import notation, we
weren't using the parsed our variable in the
call_subunit_run call.

Also, it turns out that --pdb and -n are really
both work just fine calling call_subunit_run.  So
we can remove the def call_testtools_run and just
consolidate pdb and -n into a single statement and
runner call.

I looked at using a secondary exclusive group
between no-discover and pdb, but it it shouldn't
matter if a user tries to specify both, the or
statement should evaluate correctly and things
should just work even if it's pointless to type
in `tox -epy27 -- -n --pdb project.tests.test_name`.

Change-Id: Icd61d9fb710807a508c95e8080865429a1f34fb4
Closes-Bug: #1499891