Add colored output feature to subunit-trace
This commit enables that subunit-trace shows colored output. Because
there are some people who like colored output. While this doesn't work
for the OpenStack gate for local testing it would be nice to appease
those folks and have a --color flag that adds different colors on the
console output from subunit-trace.

Change-Id: Iac48591786d10c908d63e6df94b979ded7d8dbe1
Closes-Bug: #1519044