Add tool to create a subunit stream
This commit adds a generate_subunit.py script which is used to
generate a subunit stream with a single result for a period of time.
It takes 2 mandatory args and 2 optional to specify the start time,
stop time, and optionally the status (it defaults to success) and an
id for test (if one isn't provided 'devstack' is used) The resulting
stream is written to STDOUT. There is some overlap with subunit-output
from tool from python-subunit, but this is a much smaller scope to just
just do a test_id, status, and timestamps. subunit-output doesn't support
timestamps. Eventually it'll be good to add the missing pieces to
subunit-output at which point we can likely deprecate and remove this.

The intent here is to leverage this to inject 'test results' into the
subunit2sql db to reflect failures that occur before tempest (or any
other test suite) is run. This is necessary for the openstack-health
dashboard. (otherwise it makes our failure rate look much better than
it is) This is only needed until we get a zuul mysql reporter in place
which can give us the higher level run information.

Change-Id: Icc7df33e4d73ba6322af38fbdf3aea230f2fcf4d