Remove the TODO file
The problem with TODO files is they require constant commits to keep
up to date, and this rarely happens. As is the case here where this
file is now quite stale and hasn't ever been updated. This commit
just drops the file in favor of just using launchpad as our task
tracking solution.

Change-Id: Id576d0f4dab2ecaf28db93909b7860306898a759