Construct a list of test cases instead of passing a regexp
The way how we handled the regular expressions had a lot of
limitation.

 - We are not able to pass huge arguments to testr, it makes
   difficult to have long list if accepted and/or rejected test cases,
   but we can pass a path to a file of test cases,
   which can be arbitrary big.
 - How we wanted to handle the backlists before was not worked together
   with the regular selecting regex because it consumed the pattern.
   Now the blacklisting happens in a separated phase after the selecting
   regex search.

This change just allows the new code path to run when both
 a blacklist_file and a selecting regexp specified.

The new way depends on the usual test discovery and just
filters the output of the discovery command,
this strategy can be the default in the future, now I just
wanted to preserve the old behavior as much as possible in
all the other cases.

Change-Id: Ie8e5928e286d0c9076c4eee23319149c9869a6fa
Closes-Bug: #1506215