Handle overlapping black regexes
It's very common for blacklist files and/or a black regex to have
overlapping matches. In this case ostestr was incorrectly trying to
remove tests from the set of test cases more than once. This commit
fixes this case by simply adding a check if the test is in the set
before trying to delete it. If it's not there, then something else
already removed it from the set of tests we'll be running so we can
just move on.

Change-Id: I1dabf9fb9c182af3dd6e124e79e54e1eb99bed82