Track failures during sorting of test results
The current logic only considers the first occurrence of a test class
and doesn't update a class's pass/fail status after that. So a test
class, for example:

 TestClass
   test_1 pass
   test_2 fail
   test_3 pass

is counted as a passing test class even though it has a failure.

This adds failure tracking to the test result sorting so that any
failure in a test class will make it considered a failing test class
and be sorted at the top of the HTML result page.

Closes-Bug: #1640889

Change-Id: I06919336a09c4afda8ec3a9e7d64d305fbd169c3