Fix .testr.conf detection: test path follows discover
Apparently if -t is not specified, the path to the unit test directory
follows immediately the "discover" keyword.
This fixes the discovery for the .testr.conf in the sahara repository
(even if is being replaced by .stestr.conf, the fix may still be
useful).

Also, split the code which discover the legacy values in its own function
to simplify the testing.

Finally, disable hacking rule H405 which kept flagging incorrectly a long
multiline string as docstring.

Change-Id: Ide155a8e6b2b746c81388bacc0822c68d853b5a1